Fixed the HighScore page.
A pop-up now comes up asking the user to enter a new Highscore. Part of ConcentrationGame

Also added saving for the number of cards in the game.

Scores are then put in order when a new score is saved. If you are just viewing the score the scores are not reshuffled.

Removed some unused classes as well.